perf(console): keep trace recovery off startup path
