+++ backlog/tasks/task-23113.5 - Preserve-trace-lineage-across-forks-and-temporary-chat-promotion.md
@@ -59,7 +59,7 @@ Reason: This task implements ADR-097's durable shared-prefix ownership and tempo
 - Preserved trace turn aliases through temporary forks, durable commits, restart hydration, and nested forks with version-1 metadata compatibility.
 - Made temporary Capture On pause before provider entry with Save & Send, explicit Capture Off, and Cancel actions. Save & Send promotes only the existing conversation, leaves the unsent USER leaf for durable turn acceptance, refreshes durable Library authority, and cannot duplicate the prompt or enter the provider without trace admission.
 - Addressed independent-review blockers covering production capture policy freezing, stale temporary authority, restored/nested lineage, and duplicate ordinary conversation rows.
-- Kept the trace fork boundary annotations off the UI-ready import path; the required module-census guardrail passes at 971/972 modules after the change.
+- Kept trace fork annotations off the runtime import path and consolidated the always-mounted recovery callouts so the compatibility module stays off the UI-ready path; the required module-census guardrail passes locally at 970/972 modules after the change.
 - Verified after rebasing onto `origin/dev`: 905 focused tests passed with 2 expected loopback-permission skips; Ruff, compileall, and `git diff --check` passed. Two unrelated hydration tests and one atomic-promotion test were separately reproduced unchanged on the baseline `origin/dev` worktree.
 - ADR required: yes; implemented `backlog/decisions/097-console-reference-backed-semantic-trace-ledger.md` without changing its ownership model.
 <!-- SECTION:NOTES:END -->
